requirements.txt
aiohttp==3.6.3
aioresponses==0.6.4
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
httpretty==1.0.2
idna==2.10
idna-ssl==1.1.0
lxml==4.5.2
multidict==4.7.6
pkg-resources==0.0.0
python-Wappalyzer==0.3.1
requests==2.24.0
soupsieve==2.0.1
typing-extensions==3.7.4.3
urllib3==1.25.10
yarl==1.5.1
